ai_service/requirements.txt
pillow
flask
flask-cors
easyocr
opencv-python
pytesseract